+++ app/Http/Controllers/Admin/RoleController.php
@@ -14,17 +14,17 @@ class RoleController extends Controller
 {
     function show()
     {
-        // if (!Gate::allows('role.show')) {
-        //     abort(403);
-        // }
+        if (!Gate::allows('role.show')) {
+            abort(403);
+        }
         $roles = Role::all();
         return view('admin.role.show', compact('roles'));
     }
     function add()
     {
-        // if (!Gate::allows('role.add')) {
-        //     abort(403);
-        // }
+        if (!Gate::allows('role.add')) {
+            abort(403);
+        }
         $permissions = Permission::all()
             ->groupBy(function ($permission) {
                 return explode('.', $permission->slug)[0];
@@ -33,9 +33,9 @@ class RoleController extends Controller
     }
     function store(Request $request)
     {
-        // if (!Gate::allows('role.add')) {
-        //     abort(403);
-        // }
+        if (!Gate::allows('role.add')) {
+            abort(403);
+        }
         $request->validate(
             [
                 'name' => ['required'],
@@ -50,9 +50,9 @@ class RoleController extends Controller
     }
     function delete($id)
     {
-        // if (!Gate::allows('role.delete')) {
-        //     abort(403);
-        // }
+        if (!Gate::allows('role.delete')) {
+            abort(403);
+        }
         $role_permission = Role_Permission::all()->where('role_id', $id);
         foreach ($role_permission as $item) {
             $item->delete();
@@ -66,9 +66,9 @@ class RoleController extends Controller
     }
     function handleAction(Request $request)
     {
-        // if (!Gate::allows('role.delete')) {
-        //     abort(403);
-        // }
+        if (!Gate::allows('role.delete')) {
+            abort(403);
+        }
         if (!isset($request->checkItem)) {
             return redirect('admin/role')->with('danger', 'You have not selected anything yet');
         }
@@ -92,9 +92,9 @@ class RoleController extends Controller
     }
     function edit($id)
     {
-        // if (!Gate::allows('role.edit')) {
-        //     abort(403);
-        // }
+        if (!Gate::allows('role.edit')) {
+            abort(403);
+        }
         $permissions = Permission::all()
             ->groupBy(function ($permission) {
                 return explode('.', $permission->slug)[0];
@@ -104,9 +104,9 @@ class RoleController extends Controller
     }
     function update(Request $request, $id)
     {
-        // if (!Gate::allows('role.edit')) {
-        //     abort(403);
-        // }
+        if (!Gate::allows('role.edit')) {
+            abort(403);
+        }
         $request->validate(
             [
                 'name' => ['required'],
